refactoring

+++ project/BooksStore.Web/Models/Services/CategoryViewModelService.cs
@@ -31,12 +31,22 @@ namespace BooksStore.Web.Models.Managers
 
         public async Task<IEnumerable<CategoryViewModel>> GetCategories(int pageNum)
         {
+            if(pageNum <= 0)
+            {
+                throw new ArgumentException("Номер страницы не может быть равен или меньше нуля");
+            }
+
             var pageSize = PageSizes.Categories;
             return _mapper.Map<IEnumerable<CategoryViewModel>>(await _categoryService.GetCategories((pageNum - 1) * pageSize, pageSize));
         }
 
         public async Task<CategoryViewModel> GetCategoryById(int categoryId)
         {
+            if (categoryId <= 0)
+            {
+                throw new ArgumentException("Id не может быть равен или меньше нуля");
+            }
+
             return _mapper.Map<CategoryViewModel>(await _categoryService.GetCategoryById(categoryId));
         }
 
@@ -47,6 +57,11 @@ namespace BooksStore.Web.Models.Managers
 
         public async Task RemoveCategoryAsync(int categoryId)
         {
+            if (categoryId <= 0)
+            {
+                throw new ArgumentException("Id не может быть равен или меньше нуля");
+            }
+
             await _categoryService.RemoveCategoryAsync(categoryId);
         }
